fix: airdrop dm message incorrect amount
